requirements.txt
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
huggingface_hub>=0.23.0
h5py>=3.10.0
numpy>=1.24.0
pyyaml>=6.0
pillow>=10.0.0
einops>=0.7.0
wandb>=0.16.0
tqdm>=4.66.0
